+++ crates/web-sys/src/features/gen_Gpu.rs
@@ -51,7 +51,7 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
     #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
-    pub fn request_adapter(this: &Gpu) -> ::js_sys::Promise<::js_sys::JsOption<GpuAdapter>>;
+    pub fn request_adapter(this: &Gpu) -> ::js_sys::Promise<::js_sys::JsNullable<GpuAdapter>>;
     #[cfg(web_sys_unstable_apis)]
     #[cfg(all(feature = "GpuAdapter", feature = "GpuRequestAdapterOptions",))]
     #[wasm_bindgen(method, js_class = "GPU", js_name = "requestAdapter")]
@@ -66,5 +66,5 @@ extern "C" {
     pub fn request_adapter_with_options(
         this: &Gpu,
         options: &GpuRequestAdapterOptions,
-    ) -> ::js_sys::Promise<::js_sys::JsOption<GpuAdapter>>;
+    ) -> ::js_sys::Promise<::js_sys::JsNullable<GpuAdapter>>;
 }